+++ client/containers/Project/Interface/InterfaceCol/InterfaceColContent.js
@@ -25,6 +25,7 @@ import { initCrossRequest } from 'client/components/Postman/CheckCrossInstall.js
 import produce from 'immer';
 import {InsertCodeMap} from 'client/components/Postman/Postman.js'
 
+const plugin = require('client/plugin.js');
 const {
   handleParams,
   crossRequest,
@@ -174,7 +175,7 @@ class InterfaceColContent extends Component {
     const params = this.props.match.params;
     const { actionId } = params;
     this.currColId = currColId = +actionId || result.payload.data.data[0]._id;
-    this.props.history.push('/project/' + params.id + '/interface/col/' + currColId);
+    // this.props.history.push('/project/' + params.id + '/interface/col/' + currColId);
     if (currColId && currColId != 0) {
       await this.handleColIdChange(currColId)
     }
@@ -311,6 +312,13 @@ class InterfaceColContent extends Component {
       validRes: []
     };
 
+    await plugin.emitHook('before_col_request', Object.assign({}, options, {
+      type: 'col',
+      caseId: options.caseId,
+      projectId: interfaceData.project_id,
+      interfaceId: interfaceData.interface_id
+    }));
+
     try {
       let data = await crossRequest(options, interfaceData.pre_script, interfaceData.after_script, createContext(
         this.props.curUid,
@@ -327,6 +335,13 @@ class InterfaceColContent extends Component {
         status: data.res.status,
         statusText: data.res.statusText
       };
+
+      await plugin.emitHook('after_col_request', result, {
+        type: 'col',
+        caseId: options.caseId,
+        projectId: interfaceData.project_id,
+        interfaceId: interfaceData.interface_id
+      });
 
       if (options.data && typeof options.data === 'object') {
         requestParams = {
@@ -468,7 +483,7 @@ class InterfaceColContent extends Component {
   async componentWillReceiveProps(nextProps) {
     let newColId = !isNaN(nextProps.match.params.actionId) ? +nextProps.match.params.actionId : 0;
 
-    if ((newColId && this.currColId && newColId !== this.currColId) || nextProps.isRander) {
+    if (newColId && ((this.currColId && newColId !== this.currColId) || nextProps.isRander)) {
       this.currColId = newColId;
       this.handleColIdChange(newColId)
     }
@@ -683,7 +698,7 @@ class InterfaceColContent extends Component {
                       {' '}
                       每个用例都有唯一的key，用于获取所匹配接口的响应数据，例如使用{' '}
                       <a
-                        href="https://yapi.ymfe.org/documents/case.html#%E7%AC%AC%E4%BA%8C%E6%AD%A5%EF%BC%8C%E7%BC%96%E8%BE%91%E6%B5%8B%E8%AF%95%E7%94%A8%E4%BE%8B"
+                        href="https://hellosean1025.github.io/yapi/documents/case.html#%E7%AC%AC%E4%BA%8C%E6%AD%A5%EF%BC%8C%E7%BC%96%E8%BE%91%E6%B5%8B%E8%AF%95%E7%94%A8%E4%BE%8B"
                         className="link-tooltip"
                         target="blank"
                       >
@@ -1002,7 +1017,7 @@ class InterfaceColContent extends Component {
               测试集合&nbsp;<a
                 target="_blank"
                 rel="noopener noreferrer"
-                href="https://yapi.ymfe.org/documents/case.html"
+                href="https://hellosean1025.github.io/yapi/documents/case.html"
               >
                 <Tooltip title="点击查看文档">
                   <Icon type="question-circle-o" />
@@ -1172,8 +1187,8 @@ class InterfaceColContent extends Component {
             </Row>
             <Row type="flex" justify="space-around" className="row" align="middle">
               <Col span={3} className="label">
-                邮件通知
-                <Tooltip title={'测试不通过时，会给项目组成员发送邮件'}>
+                消息通知
+                <Tooltip title={'测试不通过时，会给项目组成员发送消息通知'}>
                   <Icon
                     type="question-circle-o"
                     style={{